+++ NEWS.md
@@ -6,6 +6,7 @@
 - Improve the managing of POST failures.
 - Improve the handling of cached files.
 - Add the POST request for .tar.gz archive.
+- Change from copy to move when adding files to cache after zip download.
 - Centralize retry pattern.
 - Expand tests.
 - Add GHA.